requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Google API
google-auth==2.23.4
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Slack SDK
slack-sdk==3.26.1
slack-bolt==1.18.0

# Machine Learning
scikit-learn==1.3.2
numpy==1.26.2

# NLP
mecab-python3==1.0.6

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
python-multipart==0.0.6

# Async & Scheduling
apscheduler==3.10.4

# HTTP Client
httpx==0.25.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
